backend/requirements.txt
# FastAPI
fastapi==0.73.0
python-multipart==0.0.5
uvicorn[standard]==0.17.4
watchgod==0.7
httpx==0.22.0
pymediawiki==0.7.1
mwparserfromhell==0.6.4